uc-theme: Rename "blogs" module back to "blog", update references
