+++ packages/uc-site/src/pages/blogs/[id].astro
@@ -1,7 +1,7 @@
 ---
 import { getCollection, render, getEntries } from "astro:content";
 import { PageLayout } from "uc-theme/components";
-import { getBlogPublishedAt } from "../../../../uc-theme/src/blogs";
+import { getBlogPublishedAt } from "uc-theme/blog";
 
 export async function getStaticPaths() {
   const posts = await getCollection("blog");
